backend/requirements.txt
pymupdf==1.23.3
easyocr==1.7.1
pillow==10.0.0
